+++ tablet_qt/tasks/isaaqed.h
@@ -40,6 +40,9 @@ public:
     virtual QString shortname() const override;
     virtual QString longname() const override;
     virtual QString description() const override;
+    virtual TaskImplementationType implementationType() const override {
+        return TaskImplementationType::UpgradableSkeleton;
+    }
 
 public:
     static const QString ISAAQED_TABLENAME;